# swagger/lecture.search.yaml
openapi: 3.0.0
info:
  title: 강의 검색 API
  description: 사용자가 입력한 키워드로 강의 검색이 가능한 API
  version: 1.0.0
paths:
  /api/lectures/search:
    get:
      summary: 강의 검색
      description: |
        사용자가 입력한 검색어 기준으로 강의 목록을 검색
      tags:
        - 강의 검색
      parameters:
        - in: query
          name: keyword
          description: 검색 키워드
          required: true
          schema:
            type: string
            example: "JavaScript"
      responses:
        "200":
          description: 강의 검색 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  isSuccess:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: 강의 검색 성공
                  result:
                    type: array
                    description: 검색 결과 목록
                    items:
                      type: object
                      properties:
                        lectureId:
                          type: integer
                          description: 강의 ID
                          example: 1
                        lectureName:
                          type: string
                          description: 강의명
                          example: JavaScript 입문 강의
                        instructorName:
                          type: string
                          description: 강사명
                          example: 김강사
                        platform:
                          type: string
                          description: 플랫폼명
                          example: 클래스101
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  isSuccess:
                    type: boolean
                    example: false
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: Invalid request